+++ frontend/src/views/Home.vue
@@ -3,7 +3,7 @@
 		<template #body>
 			<div class="flex flex-col items-center mt-5 p-4 gap-7">
 				<QuickLinks :items="quickLinks" title="Quick Links" />
-				<RequestsPanel />
+				<RequestPanel />
 			</div>
 		</template>
 	</BaseLayout>
@@ -13,7 +13,7 @@
 
 import QuickLinks from "@/components/QuickLinks.vue"
 import BaseLayout from "@/components/BaseLayout.vue"
-import RequestsPanel from "@/components/RequestsPanel.vue"
+import RequestPanel from "@/components/RequestPanel.vue"
 
 import { inject } from "vue"
